Removed Nitrox.Bootloader project
Found a way to use NitroxPatcher as entrypoint while still being able to resolve DLL dependencies dynamically from either Nitrox Launcher location or game location.
