--- Nitrox.Bootloader/Nitrox.Bootloader.csproj
@@ -1,8 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net472</TargetFramework>
-        <Nullable>disable</Nullable>
-    </PropertyGroup>
-
-</Project>
